Change it to sort by release date instead of title
This will take any (non-smart) playlist and sort the items by the airdate. For example, if you want Wrestling Shows & Events/PPVs to be in the correct air order.

This will take the playlist and sort the weekly shows and Events/PPVs into the order they were aired on the same playlist. You can leave them in separate Movie/TV libraries and this will handle the order.